refactor: re-organize modules
