+++ src/main.js
@@ -1,4 +1,4 @@
-import { pollPlaylistReady } from "./library";
+import { pollPlaylistReady } from "./library/index";
 import "./main.css";
 
 // Entry-point
